Refactor TabContainer import paths in AskBookmarkSettings and MemorizeDocument components
