--- app/bibleview-js/src/components/tabs/index.ts
@@ -1,21 +1,0 @@
-/*
- - Copyright (c) 2021-2022 Martin Denham, Tuomas Airaksinen and the AndBible contributors.
- -
- - This file is part of AndBible: Bible Study (http://github.com/AndBible/and-bible).
- -
- - AndBible is free software: you can redistribute it and/or modify it under the
- - terms of the GNU General Public License as published by the Free Software Foundation,
- - either version 3 of the License, or (at your option) any later version.
- -
- - AndBible is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- - without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
- - See the GNU General Public License for more details.
- -
- - You should have received a copy of the GNU General Public License along with AndBible.
- - If not, see http://www.gnu.org/licenses/.
- */
-
-// Export all tab components for easy importing
-export { default as TabContainer } from './TabContainer.vue';
-export { default as TabNavigation } from './TabNavigation.vue';
-export { default as TabPanel } from './TabPanel.vue';
